+++ R/select.with.strikes.r
@@ -8,7 +8,8 @@
 #' @param accuracy Accuracy of parties' peremptory strikes; a number between 0 and 1; default value is .15.
 #' @return A vector of probabilities for 0:jury_n initial guilty votes
 #' @description Calculates and returns probability distribution of initial votes for guilty verdict
-#'              from 0:jury_n with options for peremptory strikes and strike accuracy
+#'              from 0:jury_n with options for peremptory strikes and strike accuracy. To select jury
+#'              without strikes, keep pstrikes=0 and dstrikes=0.
 #' @examples
 #'    library(sate)
 #'    select.with.strikes(p_g=.70, jury_n=6)